test: tricky no-extras machinery
The latest "responses" package installs PyYaml for its own use.  It
fails if we unintall PyYaml.  So now we let it import it, but then we
remove it from sys.modules so our import will fail.

Added tests to make sure it's all working the way we need it to.