Fix the invalid `scriv.user_nick` setting, which git rejects
git rejects the `scriv.user_nick` setting because of the underscore.

If specified in a command like:

```
git config get scriv.user_nick
```

git will reject the name as invalid and fail with exit code 1.

If present in a git config file, git will fail to parse the file
and fail with exit code 128.

The git reference manual has documented that only alphanumeric
and hyphen characters are allowed since at least git 2.0.5,
which was released in 2014 (11 years ago at the time of writing).

https://git-scm.com/docs/git-config/2.0.5#_configuration_file

No attempt has been made to provide backward compatibility
with the `scriv.user_nick` setting (with an underscore)
because it has been impossible for users to use the key name.

Fixes #130